+++ auditlog_tests/admin.py
@@ -1,6 +1,7 @@
 from django.contrib import admin
 
-from auditlog.registry import auditlog
+from auditlog.registry import get_default_auditlogs
 
-for model in auditlog.get_models():
-    admin.site.register(model)
+for auditlog in get_default_auditlogs():
+    for model in auditlog.get_models():
+        admin.site.register(model)