+++ auditlog_tests/test_registry.py
@@ -1,0 +1,16 @@
+from auditlog.registry import AuditlogModelRegistry
+
+create_only_auditlog = AuditlogModelRegistry(
+    create=True,
+    update=False,
+    delete=False,
+    access=False,
+    m2m=False,
+)
+update_only_auditlog = AuditlogModelRegistry(
+    create=False,
+    update=True,
+    delete=False,
+    access=False,
+    m2m=False,
+)